OCPBUGSM-29621 Unify client Scheme registration
The controller client, spokeClient, and tests client all register the
runtime.Scheme differently. This has led to bugs like the one linked to
this commit where the controller tries to interact with a custom
resource but doesn't have the right Scheme registered.

This commit uifies the process used to register these Schemes so that
the same one used everywhere. This prevents having schemes registered
for tests that are not registered in the real client and the other way
around.
